feat: Move Node Builder types to alethia-reth-node
