+++ crates/consensus/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod builder;
 pub mod eip4396;
 pub mod validation;
