--- crates/network/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "alethia-reth-network"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-
-[lib]
-path = "src/lib.rs"
-
-[features]
-default = []
-
-[dependencies]
-alethia-reth-chainspec = { path = "../chainspec" }
-eyre = { workspace = true }
-reth = { workspace = true }
-reth-ethereum = { workspace = true }
-reth-node-api = { workspace = true }
-reth-node-builder = { workspace = true }
-tracing = { workspace = true }
