+++ crates/node/src/builder/consensus.rs
@@ -1,12 +1,11 @@
 use std::sync::Arc;
 
+use alethia_reth_chainspec::spec::TaikoChainSpec;
+use alethia_reth_consensus::validation::TaikoBeaconConsensus;
+use alethia_reth_primitives::engine::TaikoEngineTypes;
 use reth_ethereum::EthPrimitives;
 use reth_node_api::{FullNodeTypes, NodeTypes};
 use reth_node_builder::{BuilderContext, components::ConsensusBuilder};
-
-use crate::validation::TaikoBeaconConsensus;
-use alethia_reth_chainspec::spec::TaikoChainSpec;
-use alethia_reth_primitives::engine::TaikoEngineTypes;
 
 /// A basic Taiko consensus builder.
 #[derive(Debug, Default, Clone)]
