+++ crates/rpc/src/eth/mod.rs
@@ -1,5 +1,4 @@
 pub mod auth;
-pub mod builder;
 pub mod error;
 #[allow(clippy::module_inception)]
 pub mod eth;
